pyproject.toml
[tool.poetry]
name = "yatdlm"
version = "0.1.0"
description = ""
authors = ["Kaci Adjou <adjou.kaci@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.12"
django = ">=5.0.2"
psycopg2-binary = "^2.8.5"
pydantic = "^2.3"

[tool.poetry.dev-dependencies]
ipython = "^8.3.0"
pylint = ">=2.5.3"
django-debug-toolbar = "^3.4.0"

[tool.poetry.group.dev.dependencies]
factory-boy = "^3.2.1"
black = "^23.1.0"
flake8 = "^6.0.0"

[flake8]
max-line-length = 90

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
